♻️ refactor(web): add useProfileSocket
