+++ apps/web/src/ui/auth/widgets/pages/SignUp/index.tsx
@@ -4,6 +4,7 @@ import { useRest } from '@/ui/global/hooks/useRest'
 import { useAuthContext } from '@/ui/auth/contexts/AuthContext'
 import { useSignUpPage } from './useSignUpPage'
 import { SignUpPageView } from './SignUpPageView'
+import { useProfileSocket } from '@/ui/global/hooks/useProfileSocket'
 
 export const SignUpPage = () => {
   const { authService, profileService } = useRest()
@@ -14,7 +15,9 @@ export const SignUpPage = () => {
     isResendingEmail,
     handleFormSubmit,
     handleResendEmail,
+    handleUserCreated,
   } = useSignUpPage(authService, user !== null)
+  useProfileSocket(handleUserCreated)
 
   return (
     <SignUpPageView
